Refactor interceptor apply out of the interceptor list
