+++ hearth/lib/hearth/middleware/build.rb
@@ -18,7 +18,8 @@ module Hearth
       # @param context
       # @return [Output]
       def call(input, context)
-        interceptor_error = context.interceptors.apply(
+        interceptor_error = Interceptor.apply(
+          interceptors: context.interceptors,
           hook: Interceptor::Hooks::MODIFY_BEFORE_SERIALIZATION,
           input: input,
           context: context,
@@ -27,7 +28,8 @@ module Hearth
         )
         return Hearth::Output.new(error: interceptor_error) if interceptor_error
 
-        interceptor_error = context.interceptors.apply(
+        interceptor_error = Interceptor.apply(
+          interceptors: context.interceptors,
           hook: Interceptor::Hooks::READ_BEFORE_SERIALIZATION,
           input: input,
           context: context,
