+++ hearth/lib/hearth/middleware/sign.rb
@@ -14,7 +14,8 @@ module Hearth
       # @param context
       # @return [Output]
       def call(input, context)
-        interceptor_error = context.interceptors.apply(
+        interceptor_error = Interceptor.apply(
+          interceptors: context.interceptors,
           hook: Interceptor::Hooks::MODIFY_BEFORE_SIGNING,
           input: input,
           context: context,
@@ -23,7 +24,8 @@ module Hearth
         )
         return Hearth::Output.new(error: interceptor_error) if interceptor_error
 
-        interceptor_error = context.interceptors.apply(
+        interceptor_error = Interceptor.apply(
+          interceptors: context.interceptors,
           hook: Interceptor::Hooks::READ_BEFORE_SIGNING,
           input: input,
           context: context,
@@ -35,7 +37,8 @@ module Hearth
         sign_request(context)
         output = @app.call(input, context)
 
-        interceptor_error = context.interceptors.apply(
+        interceptor_error = Interceptor.apply(
+          interceptors: context.interceptors,
           hook: Interceptor::Hooks::READ_AFTER_SIGNING,
           input: input,
           context: context,
