+++ hearth/spec/hearth/middleware/initialize_spec.rb
@@ -5,27 +5,41 @@ module Hearth
     describe Initialize do
       let(:app) { double('app') }
       let(:input) { double('Type::OperationInput') }
-      let(:interceptors) { double('interceptors') }
-      let(:context) { double('context', interceptors: interceptors) }
+
+      let(:request) { double('request') }
+      let(:response) { double('response') }
+      let(:interceptors) { double('interceptors', each: []) }
+      let(:context) do
+        Context.new(
+          request: request,
+          response: response,
+          interceptors: interceptors
+        )
+      end
 
       subject { Initialize.new(app) }
 
       describe '#call' do
         context 'no errors' do
           it 'calls all of the interceptor hooks and the app' do
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::READ_BEFORE_EXECUTION
-                    ))
-            expect(interceptors).to receive(:apply)
+                    )).ordered
+
+            expect(app).to receive(:call).ordered
+
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::MODIFY_BEFORE_COMPLETION
-                    ))
-            expect(interceptors).to receive(:apply)
+                    )).ordered
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::READ_AFTER_EXECUTION
-                    ))
-            expect(app).to receive(:call)
+                    )).ordered
 
             subject.call(input, context)
           end
@@ -34,18 +48,21 @@ module Hearth
         context 'read_before_execution error' do
           let(:error) { StandardError.new }
           it 'calls all interceptor hooks but does not call the app' do
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::READ_BEFORE_EXECUTION
                     ))
               .and_return(error)
 
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::MODIFY_BEFORE_COMPLETION
                     ))
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::READ_AFTER_EXECUTION
                     ))
             expect(app).not_to receive(:call)
