+++ hearth/spec/hearth/middleware/parse_spec.rb
@@ -17,7 +17,7 @@ module Hearth
         let(:output) { Output.new(metadata: metadata) }
         let(:request) { double('request') }
         let(:response) { double('response') }
-        let(:interceptors) { double('interceptors', apply: nil) }
+        let(:interceptors) { double('interceptors', each: []) }
         let(:context) do
           Context.new(
             request: request,
@@ -40,20 +40,23 @@ module Hearth
 
         it 'calls all of the interceptor hooks after the next middleware' do
           expect(app).to receive(:call).ordered
-          expect(interceptors).to receive(:apply)
+          expect(Interceptor).to receive(:apply)
             .with(hash_including(
+                    interceptors: interceptors,
                     hook: Interceptor::Hooks::MODIFY_BEFORE_DESERIALIZATION
                   )).ordered
-          expect(interceptors).to receive(:apply)
+          expect(Interceptor).to receive(:apply)
             .with(hash_including(
+                    interceptors: interceptors,
                     hook: Interceptor::Hooks::READ_BEFORE_DESERIALIZATION
                   )).ordered
 
           expect(error_parser).to receive(:parse).ordered
           expect(data_parser).to receive(:parse).ordered
 
-          expect(interceptors).to receive(:apply)
+          expect(Interceptor).to receive(:apply)
             .with(hash_including(
+                    interceptors: interceptors,
                     hook: Interceptor::Hooks::READ_AFTER_DESERIALIZATION
                   )).ordered
 
@@ -111,8 +114,9 @@ module Hearth
           it 'returns output with the error set' do
             expect(app).to receive(:call)
               .and_return(output)
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::MODIFY_BEFORE_DESERIALIZATION
                     )).and_return(error)
 
@@ -130,8 +134,14 @@ module Hearth
           it 'returns output with the error set' do
             expect(app).to receive(:call)
               .and_return(output)
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
+                      hook: Interceptor::Hooks::MODIFY_BEFORE_DESERIALIZATION
+                    ))
+            expect(Interceptor).to receive(:apply)
+              .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::READ_BEFORE_DESERIALIZATION
                     )).and_return(error)
 
@@ -149,8 +159,19 @@ module Hearth
           it 'returns output with the error set' do
             expect(app).to receive(:call)
               .and_return(output)
-            expect(interceptors).to receive(:apply)
+            expect(Interceptor).to receive(:apply)
               .with(hash_including(
+                      interceptors: interceptors,
+                      hook: Interceptor::Hooks::MODIFY_BEFORE_DESERIALIZATION
+                    ))
+            expect(Interceptor).to receive(:apply)
+              .with(hash_including(
+                      interceptors: interceptors,
+                      hook: Interceptor::Hooks::READ_BEFORE_DESERIALIZATION
+                    ))
+            expect(Interceptor).to receive(:apply)
+              .with(hash_including(
+                      interceptors: interceptors,
                       hook: Interceptor::Hooks::READ_AFTER_DESERIALIZATION
                     )).and_return(error)
